refactor(agent-tars-cli): decouple
