+++ multimodal/agent-interface/src/index.ts
@@ -7,6 +7,7 @@ export * from './agent';
 export * from './agent-options';
 export * from './agent-run-options';
 export * from './agent-instance';
+export * from './agent-constructor';
 export * from './tool';
 export * from './tool-call-engine';
 export * from './agent-event-stream';
